requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-groq
langchain-huggingface
langchain-qdrant
qdrant-client
sentence-transformers
fastapi
uvicorn
pypdf
python-dotenv